v1/galion/services/cms-service/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Email validation
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# Testing (optional but recommended)
pytest==7.4.3
httpx==0.25.2

